requirements.txt
fastapi[standard]==0.121.3
python-dotenv==1.2.1
beanie==2.0.1